requirements.txt
altgraph==0.17.4
asgiref==3.6.0
attrs==23.1.0
awscli==1.29.12
awsebcli==3.20.10
bcrypt==4.0.1
beautifulsoup4==4.12.2
botocore==1.29.158
cement==2.8.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.3
cryptography==41.0.2
distlib==0.3.7
Django==4.2.7
django-cors-headers==4.0.0
django-debug-toolbar==4.0.0
django-pyodbc-azure==2.1.0.0
djangorestframework==3.14.0
docutils==0.16
et-xmlfile==1.1.0
exceptiongroup==1.1.2
filelock==3.12.2
greenlet==3.0.1
h11==0.14.0
idna==3.4
install==1.3.5
jmespath==1.0.1
lxml==4.9.3
messagebox==0.1.0
numpy==1.26.2
O365==2.0.27
oauthlib==3.2.2
openpyxl==3.1.2
outcome==1.2.0
packaging==23.1
pandas==2.1.3
pathspec==0.10.1
pefile==2023.2.7
platformdirs==3.9.1
psycopg2==2.9.6
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.8
PyJWT==2.8.0
pymssql==2.2.10
pyodbc==5.0.1
pyOutlook==4.2.2
pypiwin32==223
pypyodbc==1.3.6
PySocks==1.7.1
pyspnego==0.9.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==5.3.1
requests==2.31.0
requests-ntlm==1.2.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rsa==4.7.2
RuNAs==0.1
s3transfer==0.6.1
selenium==4.10.0
semantic-version==2.8.5
SharePlum==0.5.1
six==1.14.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.23
sqlparse==0.4.4
stringcase==1.2.0
termcolor==1.1.0
tqdm==4.65.0
trio==0.22.2
trio-websocket==0.10.3
typer==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==1.26.16
virtualenv==20.24.2
wcwidth==0.1.9
webdriver-manager==3.8.6
wsproto==1.2.0
